fixing blossom implmentation
